+++ src/siun/main.py
@@ -4,7 +4,6 @@
 
 import datetime
 from pathlib import Path
-from typing import Any
 
 import click
 
@@ -18,7 +17,7 @@ from siun.errors import (
     SiunStateUpdateError,
 )
 from siun.formatting import Formatter, OutputFormat
-from siun.models import V2Threshold
+from siun.models import V2Criterion, V2Threshold
 from siun.notification import INSTALLED_FEATURES as INSTALLED_NOTIFICATION_FEATURES
 from siun.state import FormatObject, Updates, load_state, update_state_with_available_packages
 
@@ -26,13 +25,15 @@ CONTEXT_SETTINGS = {"help_option_names": ["-h", "--help"]}
 INSTALLED_FEATURES: set[str] = INSTALLED_NOTIFICATION_FEATURES
 
 
-def get_formatted_state_text(format_object: FormatObject, output_format: str, custom_format: str) -> str:
+def get_formatted_state_text(format_object: FormatObject, output_format: OutputFormat, custom_format: str) -> str:
     """Generate formatted output text from update state."""
     formatter = Formatter()
     formatter_kwargs = {}
-    if output_format == OutputFormat.CUSTOM.value:
+    if output_format == OutputFormat.CUSTOM:
         formatter_kwargs["template_string"] = custom_format
-    formatted_output, format_options = getattr(formatter, f"format_{output_format}")(format_object, **formatter_kwargs)
+    formatted_output, format_options = getattr(formatter, f"format_{output_format.value}")(
+        format_object, **formatter_kwargs
+    )
     return click.style(formatted_output, **format_options)
 
 
@@ -41,7 +42,7 @@ def _get_updates(
     no_cache: bool,
     no_update: bool,
     cmd_available: str,
-    criteria: dict[str, Any],
+    criteria: list[V2Criterion],
     thresholds: list[V2Threshold],
     cache_min_age_minutes: int,
     state_file_path: Path,
@@ -138,9 +139,12 @@ def cli():  # noqa: D103 # pragma: no cover
 @click.option("--no-update", "-U", is_flag=True, show_default=True, default=False, help="Don't get updates, only check")
 @click.option("--cache/--no-cache", " /-n", show_default=True, default=True, help="Ignore existing state on disk")
 @click.option(
-    "--output-format", "-o", default="plain", type=click.Choice([of.value for of in OutputFormat], case_sensitive=False)
+    "--output-format",
+    "-o",
+    default=OutputFormat.PLAIN,
+    type=click.Choice(OutputFormat, case_sensitive=False),
 )
-def check(*, output_format: str, cache: bool, no_update: bool, quiet: bool, config_path: Path):
+def check(*, output_format: OutputFormat, cache: bool, no_update: bool, quiet: bool, config_path: Path):
     """Check for urgency of available updates."""
     if no_update and not cache:
         raise SiunCLIError(message="--no-update and --no-cache options are mutually exclusive")
@@ -157,7 +161,7 @@ def check(*, output_format: str, cache: bool, no_update: bool, quiet: bool, conf
             no_cache=not cache,
             no_update=no_update,
             cmd_available=config.cmd_available,
-            criteria=config.criteria,
+            criteria=config.v2_criteria,
             thresholds=config.sorted_thresholds,
             cache_min_age_minutes=config.cache_min_age_minutes,
             state_file_path=config.state_file,